🏷️ (Refactor button types to component.types and export from lib/types): Refactored Image types to use NextImage types: ImageProps
