+++ components/Image.tsx
@@ -1,21 +1,10 @@
-import Image from 'next/image';
+import Image, {ImageProps} from 'next/image';
 
 const customLoader = ({src, width, quality}) => {
   return `${src}?w=${width}&q=${quality || 75}`;
 };
 
-interface CustomImageProps {
-  src: string;
-  alt: string;
-  width: string | number;
-  height: string | number;
-  className?: string;
-  layout?: 'responsive' | 'fill' | 'intrinsic' | 'fixed';
-  quality?: number;
-  objectFit?: 'contain' | 'cover' | 'fill' | 'none' | 'scale-down';
-  priority?: boolean;
-}
-export default function CustomImage(props: CustomImageProps) {
+export default function CustomImage(props: ImageProps) {
   return (
     <Image
       loader={customLoader}
