+++ lib/types.ts
@@ -1,2 +1,3 @@
 export * from './types/blog.types';
 export * from './types/user.types';
+export * from './types/component.types';
